Changed DB scheme, basic structure added
