+++ update-processor/src/main/java/ru/tbank/processor/service/group/GroupChatUpdateProcessingService.java
@@ -1,4 +1,4 @@
-package ru.tbank.processor.service;
+package ru.tbank.processor.service.group;
 
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
@@ -10,7 +10,9 @@ import org.telegram.telegrambots.meta.api.objects.message.Message;
 import ru.tbank.common.entity.dto.DeletedTextMessageDTO;
 import ru.tbank.common.entity.text.TextModerationSettings;
 import ru.tbank.common.entity.text.TextProcessingResult;
-import ru.tbank.processor.service.filter.text.TextFilter;
+import ru.tbank.processor.service.ChatModerationSettingsService;
+import ru.tbank.processor.service.TelegramClientService;
+import ru.tbank.processor.service.group.filter.text.TextFilter;
 
 import java.util.Collections;
 import java.util.List;
