+++ update-processor/src/main/java/ru/tbank/processor/service/group/filter/text/impl/CustomEmojisFilter.java
@@ -1,4 +1,4 @@
-package ru.tbank.processor.service.filter.text.impl;
+package ru.tbank.processor.service.group.filter.text.impl;
 
 import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.NullMarked;
@@ -8,26 +8,30 @@ import ru.tbank.common.entity.FilterMode;
 import ru.tbank.common.entity.text.TextFilterSettings;
 import ru.tbank.common.entity.text.TextModerationSettings;
 import ru.tbank.common.entity.text.TextProcessingResult;
-import ru.tbank.processor.service.filter.text.FilterCost;
-import ru.tbank.processor.service.filter.text.TextEntityType;
-import ru.tbank.processor.service.filter.text.TextFilter;
+import ru.tbank.processor.service.TelegramClientService;
+import ru.tbank.processor.service.group.filter.text.FilterCost;
+import ru.tbank.processor.service.group.filter.text.TextEntityType;
+import ru.tbank.processor.service.group.filter.text.TextFilter;
 
 @Slf4j
 @NullMarked
 @Component
-public class LinksFilter extends TextFilter {
+public class CustomEmojisFilter extends TextFilter {
 
-    public LinksFilter() {
-        super(FilterCost.VERY_LOW);
+    private final TelegramClientService telegramClientService;
+
+    public CustomEmojisFilter(TelegramClientService telegramClientService) {
+        super(FilterCost.MEDIUM);
+        this.telegramClientService = telegramClientService;
     }
 
     @Override
     public TextProcessingResult process(Message message, TextModerationSettings moderationSettings) {
-        var filterSettings = moderationSettings.getLinksFilterSettings();
+        var filterSettings = moderationSettings.getCustomEmojisFilterSettings();
         var checkResult = filterSettings.isEnabled() && message.hasEntities()
-                && isContainsBlockedEntity(message, filterSettings, TextEntityType.URL);
+                && isContainsBlockedEntity(message, filterSettings, TextEntityType.CUSTOM_EMOJI);
 
-        return checkResult ? TextProcessingResult.LINK_FOUND : TextProcessingResult.OK;
+        return checkResult ? TextProcessingResult.TAG_FOUND : TextProcessingResult.OK;
     }
 
     @Override
@@ -36,8 +40,8 @@ public class LinksFilter extends TextFilter {
         return message.getEntities().stream()
                 .filter(entity -> entity.getType().equals(entityType.name().toLowerCase()))
                 .anyMatch(entity -> {
-                    var entityLink = entity.getText().replaceFirst("^https?://", "");
-                    var checkResult = filterSettings.getExclusions().stream().anyMatch(entityLink::startsWith);
+                    var stickerSet = telegramClientService.getEmojiPack(entity.getCustomEmojiId());
+                    var checkResult = filterSettings.getExclusions().contains(stickerSet.getFirst().getSetName());
                     return (filterSettings.getExclusionMode() == FilterMode.WHITE_LIST) != checkResult;
                 });
     }
