+++ update-processor/src/main/java/ru/tbank/processor/service/group/filter/text/impl/LinksFilter.java
@@ -1,5 +1,6 @@
-package ru.tbank.processor.service.filter.text.impl;
+package ru.tbank.processor.service.group.filter.text.impl;
 
+import lombok.extern.slf4j.Slf4j;
 import org.jspecify.annotations.NullMarked;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.message.Message;
@@ -7,25 +8,26 @@ import ru.tbank.common.entity.FilterMode;
 import ru.tbank.common.entity.text.TextFilterSettings;
 import ru.tbank.common.entity.text.TextModerationSettings;
 import ru.tbank.common.entity.text.TextProcessingResult;
-import ru.tbank.processor.service.filter.text.FilterCost;
-import ru.tbank.processor.service.filter.text.TextEntityType;
-import ru.tbank.processor.service.filter.text.TextFilter;
+import ru.tbank.processor.service.group.filter.text.FilterCost;
+import ru.tbank.processor.service.group.filter.text.TextEntityType;
+import ru.tbank.processor.service.group.filter.text.TextFilter;
 
+@Slf4j
 @NullMarked
 @Component
-public class BotCommandsFilter extends TextFilter {
+public class LinksFilter extends TextFilter {
 
-    public BotCommandsFilter() {
+    public LinksFilter() {
         super(FilterCost.VERY_LOW);
     }
 
     @Override
     public TextProcessingResult process(Message message, TextModerationSettings moderationSettings) {
-        var filterSettings = moderationSettings.getBotCommandsFilterSettings();
+        var filterSettings = moderationSettings.getLinksFilterSettings();
         var checkResult = filterSettings.isEnabled() && message.hasEntities()
-                && isContainsBlockedEntity(message, filterSettings, TextEntityType.BOT_COMMAND);
+                && isContainsBlockedEntity(message, filterSettings, TextEntityType.URL);
 
-        return checkResult ? TextProcessingResult.TAG_FOUND : TextProcessingResult.OK;
+        return checkResult ? TextProcessingResult.LINK_FOUND : TextProcessingResult.OK;
     }
 
     @Override
@@ -34,8 +36,8 @@ public class BotCommandsFilter extends TextFilter {
         return message.getEntities().stream()
                 .filter(entity -> entity.getType().equals(entityType.name().toLowerCase()))
                 .anyMatch(entity -> {
-                    var checkResult = filterSettings.getExclusions().stream()
-                            .anyMatch(exclusion -> entity.getText().startsWith(exclusion));
+                    var entityLink = entity.getText().replaceFirst("^https?://", "");
+                    var checkResult = filterSettings.getExclusions().stream().anyMatch(entityLink::startsWith);
                     return (filterSettings.getExclusionMode() == FilterMode.WHITE_LIST) != checkResult;
                 });
     }
