+++ update-processor/src/main/java/ru/tbank/processor/service/group/filter/text/impl/PhoneNumberFilter.java
@@ -1,13 +1,13 @@
-package ru.tbank.processor.service.filter.text.impl;
+package ru.tbank.processor.service.group.filter.text.impl;
 
 import org.jspecify.annotations.NullMarked;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.message.Message;
 import ru.tbank.common.entity.text.TextModerationSettings;
 import ru.tbank.common.entity.text.TextProcessingResult;
-import ru.tbank.processor.service.filter.text.FilterCost;
-import ru.tbank.processor.service.filter.text.TextEntityType;
-import ru.tbank.processor.service.filter.text.TextFilter;
+import ru.tbank.processor.service.group.filter.text.FilterCost;
+import ru.tbank.processor.service.group.filter.text.TextEntityType;
+import ru.tbank.processor.service.group.filter.text.TextFilter;
 
 @NullMarked
 @Component
