+++ update-processor/src/main/java/ru/tbank/processor/service/personal/UserRole.java
@@ -1,0 +1,6 @@
+package ru.tbank.processor.service.personal;
+
+public enum UserRole {
+    OWNER,
+    ADMIN
+}
