+++ update-receiver/src/main/resources/application.yml
@@ -15,3 +15,4 @@ spring:
 logging:
   level:
     org.apache.kafka: warn
+    ru.tbank.receiver: debug
